requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
exceptiongroup==1.2.1
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
pytest==8.2.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
tomli==2.0.1
typing-extensions==4.12.2
tzdata==2024.1
werkzeug==3.0.3
zipp==3.19.2
